remove Optional for session.reset_password
The possibility for the session namespaces to be None makes it
cumbersome to use them, because all code has to check if it is None
before accessing properties on them.

As we already initialise a default ResetPasswordNS instance in case
_reset_password, it seems like it should be OK to just say that
state.reset_password will _always_ be a ResetPasswordNS instance.

We want to initialise it on first access though, which is why we use
hasattr() (instead of initialising it to the default ResetPasswordNS
instance in the constructor) to know if it has been initialised or not.