rework most parts of how flask view returned data is formatted
I had a hard time understanding exactly how returned data was consumed
by the frontend code, so to try and improve this for the next person to
come along I added lots of documentation and questioned a lot of how
this is currently done and explained things as well as I could :).

Also, I felt that a proper data container (FluxData rather than a dict
with 'special' elements like '_status') to transport data from the
success_message/error_message functions into the MarshalWith decorator,
where it was turned into a FluxResponse would be more robust.