write data to redis only if it hasn't been modified
Use a redis transaction to not overwrite session data that has been
modified.

This is needed to safe-guard against two requests reading the session
and then making different changes to the session data. Better to fail
the second request in a detectable manner, than to randomly loose
updates to the session data.