get rid of two # type: ignore
Code in eduid-common can't know the precise type of current_app.config.

Since different apps have different current_app.config, type checkers
can't draw conclusions about properties in the confuguration. This leads
to less maintainable code, so we decided on the principle to only use
parts of current_app that can actually be expected to be the same, or
functional, throughout all apps (such as current_app.logger).

Specific parts of the apps configuration is better passed to code in
eduid-common as parameters.