add session namespace for eidas
This prevents eidas from overwriting parts of the session already used
by authn.

An actual issue was when authn initiated authentication requiring MFA,
the user chose to use Freja (eidas) and then aborted that and completed
login using a security key. When the user reached authn in this case, it
was missing it's post-authn-action it had put in the session, and
crashed.