make 'verified' return a plain list with ListElements
Re-instantiating __class__ seems to do a deep copy with Pydantic, so if
a sub-list of items was created with 'verified' and modified, it
wouldn't be reflected in the original ElementList.

This made the sub-lists somewhat useless, and the most common thing done
in the codebase was to use .to_list() on it anyways.